Bug 1886006 - Add profile avatar images. r=mossop
These SVG images have a background `<path>`, which we color using
`fill="context-fill"` in the SVG, and the `fill` property in CSS, and
they have one or more foreground image `<path>` elements, which we
color using `fill="context-stroke"` in the SVG, and the `stroke` CSS
property. (The CSS styles are applied to elements that display the
avatars, not included in this commit.)

The overall effect is a two-tone avatar image that will be dynamically
styled based on the foreground and background colors of a profile's
theme.

Note that the original Figma-exported SVGs had either black or white
fills, and a fill attribute set on the `<clipPath><rect>` found in some
images. These were modified by hand by eng, swapping in the appropriate
fill attribute value for background / foreground elements, and if a fill
was set on the clipPath rect, that attribute was removed (as it didn't
seem to do anything). The MPL license headers were auto-inserted by the
linter.

Differential Revision: https://phabricator.services.mozilla.com/D223237