Bug 1918416 - part 1 - Prevent leaking DynamicDownloadDialogBehaviors r=android-reviewers,skhan
For each new download dialog to be shown we set a new DynamicDownloadDialogBehavior.
But a previous one may now be leaked because it continues to observe store updates and
reacting to when the orientation changes.
The leak is made possible because the "download complete dialog" is always attached to
the displayed layout, only it's visibility is updated depending on if it needs to be shown.

Differential Revision: https://phabricator.services.mozilla.com/D222281