Bug 1920951 part 5 - Avoid inserting some unnecessary moves in createMoveGroupsForControlFlowEdges. r=jseward
In `createMoveGroupsForControlFlowEdges`, for at least 70% of all edges, the predecessor
and successor blocks have ranges with the same allocation. In this case `moveAtEdge` is a
no-op. We can increase this number by checking for ranges with exactly the same allocation.

For some Ion compilation workloads, this is the number of edges for which no move has to be
inserted before/after this change + the number of moves generated for control flow edges:

* Wasm PhotoShop: 86.3% to 87.1% (3112936 moves to 2918044 moves, 6% less)
* Wasm OpenOffice: 72.5% to 76.7% (2596531 moves to 2196918 moves, 15% less)
* Wasm ONNX: 79.2% to 82.9% (828700 moves to 680817 moves, 18% less)
* JS Octane: 71.3% to 76.1% (25854 moves to 21145 moves, 18% less)

This also fixes the register allocation issue reported in bug 1912154.

Differential Revision: https://phabricator.services.mozilla.com/D224192