Bug 1921706 - simplify removing libatomic on linux builds. r=ng,webrtc-reviewers DONTBUILD
Note: this produces no changes in the generated moz.build files.

o  changeset:   597701:a26d44969a83
~  user:        Michael Froman <mfroman@mozilla.com>
   date:        Mon Jun 14 17:39:29 2021 -0500
   files:       third_party/libwebrtc/build/config/linux/BUILD.gn
   description:
   Bug 1654112 - don't use libatomic on linux builds. r=ng

In moz-central a26d44969a83 I added an include of webrtc.gni for
the build_with_mozilla flag, but there is already a Mozilla
modifcation in this file that does not rely on that flag.  This
simplification will help with future google build directory move.

Differential Revision: https://phabricator.services.mozilla.com/D224069