Bug 1922149 - Make `HTMLEditor::ComputeEditingHostInternal` use common ancestor of all selection ranges r=m_kato
Selection ranges can cross editing host boundaries if no editing host has focus.
Therefore, `Selection.focusNode` may be in an editing host but there may be
no active/focused editing host.

The computation may be expensive if there are a lot of ranges and selecting
in slotted shadow tree.  However, it's rare case, so, I think it's okay for
now.

Differential Revision: https://phabricator.services.mozilla.com/D224282