Bug 1919366: Force cache domain queries for test platform exceptions on Linux, r=Jamie
These exceptions exist because of Linux platform code which can cause caching of
certain domains before the test begins. Previously, we thought that these
platform issues were 100% consistent, but it turns out that they only happen
most of the time, and can fail to happen in test verify, causing test failure.
To ensure that we don't fail in unexpected situations, call the query function
even for these Linux platform exceptions. In the best case, it won't affect
anything, and in the worst case, it'll save us from TV failure.

Differential Revision: https://phabricator.services.mozilla.com/D222880