Bug 1909391 - Do not filter out credit cards that are missing some data r=dimi
In the previous code, credit cards without names, or an expiration date,
were filtered out. But it's common that Firefox doesn't save the full
information of a card, and as a result these cards weren't showing up in
Firefox for Android, despite they were showing up in Firefox Desktop.

This patch changes the check for a credit card validity to just checking
if the number is present.

Differential Revision: https://phabricator.services.mozilla.com/D224966