Bug 1917818 - part 1 - Synchronize showing browser's bottom toolbar on top of the keyboard r=android-reviewers,sfamisa
Previously when the keyboard would be showing / hiding the entire screen will first resize to
the end height after which the keyboard will be animated.
This meant just one reflow for the engine view but a choppy layout animation.

With the new changes we'll ensure the bottom toolbar is animated in sync with the keyboard
and the engine view is resized only once
- before hiding the keyboard to ensure that it will then nicely reveal the web content
- after showing the full keyboard to ensure it gradually covered web content with no UI issue

Differential Revision: https://phabricator.services.mozilla.com/D223585