Bug 1917818 - part 2 - Improve browser layout with top toolbar updates when the keyboard is animated r=android-reviewers,skhan
Previously when the keyboard would be showing the entire screen will first resize to the end height
after which the keyboard will be animated over an empty content.
With these changes we postpone the resize to after the keyboard animation is finished to ensure a
smooth animation over the web content instead of over empty space.

Differential Revision: https://phabricator.services.mozilla.com/D223586