Bug 1807913 - Fix openLoginWebsiteInBrowserTest UI test r=ohorvath
The UI test failed yesterday on API 34 because the save logins prompt wasn't displayed.
As seen before, this might be caused by the fact that the autofill actions are performed to qucikly.
To overcome this problem, I've added a 1s wait between the autofilling actions.

The UI test successfully passed 200x on Firebase. ✅

Differential Revision: https://phabricator.services.mozilla.com/D225774