Bug 1849519 - Disable the consistency check assertion for mTrackingScripts. r=sefeng,dom-core
This is causing problems when loading real world sites in debug builds,
which is hurting automated bug finding, so disable this assertion.

mTrackingScripts look like it is only used to decide when to delay a
load sometimes, so I think it isn't a huge deal if this is inconsistent
sometimes.

Differential Revision: https://phabricator.services.mozilla.com/D224996