Bug 1917652 Part 2: Explicitly define AddRef and Release in VideoBridgeParent. r=aosmond
This change prevents VideoBridgeParent from being incorrectly reported
as leaked memory. Something about the logging used by the previous macro
confuses the leakchecker. Appropriate deallocation was confirmed by
creating manual implementations of AddRef and Release and the various
Actor lifecyle methods (not part of this patch). The change to this new
approach does the same work as the old macro, minus the conditional
logging of the old macro.

Differential Revision: https://phabricator.services.mozilla.com/D222715