Bug 1925429 [wpt PR 48688] - Merge the Safari workflows into a single workflow, a=testonly
Automatic update from web-platform-tests
Merge the Safari workflows into a single workflow

This gets rid of a lot of duplication and makes this much
simpler. This might be hard to use for PRs where we want to do
different chunking (given the matrix definition), but that is a future
challenge (one possibility is to just pass in a blob of JSON as an
input and use matrix: fromJSON()).

--

wpt-commits: b4847226b3e278f0daea1f22d137fb2da1bb50b6
wpt-pr: 48688