Bug 1877330 Part 3: Make WebGPUParent swallow unreportable device lost callbacks. r=teoxoy
This changes both the parent and the child to clarify that some device
lost callbacks from wgpu should not resolve the lost promise. It also
adds an assert to detect when we are getting the lost callback after
we've already unregistered device, which is a possible cause of timeouts
in this test.

It restores the part of the test that checks that device drop is
unobservable, which it once again is.

Differential Revision: https://phabricator.services.mozilla.com/D225733