Bug 1920821 - Enable mitmproxy 11.0.0 on Windows/MacOS r=perftest-reviewers,sparky
This patch enables performance pageload tests to now use mitmproxy
11.0.0, only on the Mac/Windows platform.
This will run on a subset of tests as the binary doesn't properly run on
old recordings (e.g. mitm 6) but seems to run fine on recordings mitm7
and greater.

Additionally, the mitmproxy version for android is bumped from 10 -> 11.
Although 10.x.x was only recently landed, 11 was released shortly after
so we might as well keep all platforms on the same version where
possible.

Differential Revision: https://phabricator.services.mozilla.com/D224383